Update association equality definition to be full diff
